requirements.txt
fastapi~=0.112.2
uvicorn==0.35.0
aiofiles~=22.1.0
aiosqlite
sqlalchemy~=2.0.41
werkzeug==3.1.3
python-dotenv~=1.1.0
pydantic~=2.11.7
numpy~=2.2.6
scikit-learn~=1.7.1
fastapi-users[sqlalchemy]